+++ src/createPlugin.js
@@ -1,0 +1,6 @@
+export default function (ruleName, rule) {
+  return {
+    ruleName,
+    rule,
+  }
+}