+++ src/formatters/stringFormatter.js
@@ -1,3 +1,4 @@
+import chalk from "chalk"
 import formatter from "postcss-reporter/lib/formatter"
 
 const minimalFormatter = formatter({
@@ -6,12 +7,26 @@ const minimalFormatter = formatter({
 })
 
 export default function (results) {
-  let output = ""
-  results.forEach(result => {
+  return results.reduce((output, result) => {
+    output += deprecationsFormatter(result.deprecations)
     output += minimalFormatter({
       messages: result.warnings,
       source: result.source,
     })
-  })
-  return output
+    return output
+  }, "")
+}
+
+function deprecationsFormatter(warnings) {
+  if (!warnings || !warnings.length) { return "" }
+
+  return warnings.reduce((output, warning) => {
+    output += chalk.yellow.bold(">> Deprecation Warning: ")
+    output += chalk.yellow(warning.text)
+    if (warning.reference) {
+      output += chalk.yellow(" See: ")
+      output += chalk.green.underline(warning.reference)
+    }
+    return output + "\n"
+  }, "\n")
 }