+++ src/postcssPlugin.js
@@ -1,128 +1,86 @@
 import postcss from "postcss"
-import rc from "rc"
+import multimatch from "multimatch"
+import { get } from "lodash"
 import path from "path"
-import resolveFrom from "resolve-from"
-import { assign, mapValues, merge, isEmpty } from "lodash"
 import { configurationError } from "./utils"
 import ruleDefinitions from "./rules"
 import disableRanges from "./disableRanges"
+import buildConfig from "./buildConfig"
+
+const numberedSeveritiesMap = new Map()
+numberedSeveritiesMap.set(0, "ignore")
+numberedSeveritiesMap.set(1, "warning")
+numberedSeveritiesMap.set(2, "error")
 
 export default postcss.plugin("stylelint", (options = {}) => {
+  const configPromise = buildConfig(options)
+
   return (root, result) => {
     // result.stylelint is the namespace for passing stylelint-related
     // configuration and data across sub-plugins via the PostCSS Result
     result.stylelint = result.stylelint || {}
     result.stylelint.ruleSeverities = {}
 
-    const config = buildConfig(options)
-
-    if (!config) {
-      throw configurationError("No configuration provided")
-    }
-    if (!config.rules) {
-      throw configurationError("No rules found within configuration")
-    }
-
-    if (config.plugins) {
-      merge(ruleDefinitions, mapValues(config.plugins, plugin => require(plugin)))
-    }
-
-    // Register details about the configuration
-    result.stylelint.quiet = config.quiet
-
-    // First check for disabled ranges, adding them to the result object
-    disableRanges(root, result)
-
-    Object.keys(config.rules).forEach(ruleName => {
-      if (!ruleDefinitions[ruleName]) {
-        throw configurationError(`Undefined rule ${ruleName}`)
+    return configPromise.then(({ config, configDir }) => {
+      if (!config) {
+        throw configurationError("No configuration provided")
       }
 
-      // If severity is 0, run nothing
-      const ruleSettings = config.rules[ruleName]
-      const ruleSeverity = (Array.isArray(ruleSettings))
-        ? ruleSettings[0]
-        : ruleSettings
-      if (ruleSeverity === 0) {
-        return
+      if (!config.rules) {
+        throw configurationError("No rules found within configuration")
       }
 
-      // Log the rule's severity
-      result.stylelint.ruleSeverities[ruleName] = ruleSeverity
+      if (config.ignoreFiles) {
+        const pathFromConfigToSource = path.relative(configDir, get(root, "source.input.file", ""))
+        if (multimatch(pathFromConfigToSource, config.ignoreFiles).length) { return }
+      }
 
-      // Run the rule with the primary and secondary options
-      ruleDefinitions[ruleName](ruleSettings[1], ruleSettings[2])(root, result)
+      if (config.numberedSeverities) {
+        warnForNumberedSeverities(result)
+      }
+
+      if (config.plugins) {
+        config.plugins.forEach(pluginPath => {
+          const plugin = require(pluginPath).default
+          ruleDefinitions[plugin.ruleName] = plugin.rule
+        })
+      }
+
+      // Register details about the configuration
+      result.stylelint.quiet = config.quiet
+
+      // First check for disabled ranges, adding them to the result object
+      disableRanges(root, result)
+
+      Object.keys(config.rules).forEach(ruleName => {
+        if (!ruleDefinitions[ruleName]) {
+          throw configurationError(`Undefined rule "${ruleName}"`)
+        }
+
+        const ruleSettings = [].concat(config.rules[ruleName])
+        const ruleSeverity = numberedSeveritiesMap.get(ruleSettings[0])
+
+        if (ruleSeverity === "ignore") { return }
+
+        // Log the rule's severity in the PostCSS result
+        result.stylelint.ruleSeverities[ruleName] = ruleSeverity
+
+        // Run the rule with the primary and secondary options
+        ruleDefinitions[ruleName](ruleSettings[1], ruleSettings[2])(root, result)
+      })
     })
   }
 })
 
-function buildConfig(options) {
-  let initialConfig
-  let configBasedir
-
-  if (options.hasOwnProperty("configFile")) {
-    const absoluteConfigPath = (path.isAbsolute(options.configFile))
-      ? options.configFile
-      : path.join(process.cwd(), options.configFile)
-    try {
-      initialConfig = require(absoluteConfigPath)
-      configBasedir = path.dirname(absoluteConfigPath)
-    } catch (e) {
-      throw configurationError(
-        `configFile path "${options.configFile}" corresponded to no file`
-      )
-    }
-  } else {
-    const passedConfig = options.hasOwnProperty("config")
-      ? options.config
-      : options
-    initialConfig = (isEmpty(passedConfig))
-      ?  rc("stylelint")
-      : passedConfig
-    configBasedir = options.configBasedir || path.dirname(initialConfig.config)
-  }
-
-  const config = extendConfig(initialConfig, configBasedir)
-
-  if (options.configOverrides) {
-    merge(config, options.configOverrides)
-  }
-
-  return config
-}
-
-function extendConfig(config, basedir = process.cwd()) {
-  // Absolutize the plugins here, because here is the place
-  // where we know the basedir for this particular config
-  const configWithAbsolutePlugins = absolutizePlugins(config, basedir)
-  if (!config.extends) return configWithAbsolutePlugins
-
-  return [].concat(config.extends).reduce((mergedConfig, extendingConfigLookup) => {
-    const extendingConfigPath = getModulePath(basedir, extendingConfigLookup)
-    const extendingConfigDir = path.dirname(extendingConfigPath)
-
-    // Now we must recursively extend the extending config
-    const extendingConfig = extendConfig(require(extendingConfigPath), extendingConfigDir)
-
-    return merge({}, extendingConfig, mergedConfig)
-  }, configWithAbsolutePlugins)
-}
-
-// Replace all plugin looksup with absolute paths
-function absolutizePlugins(config, basedir) {
-  if (!config.plugins) { return config }
-  return assign({}, config, {
-    plugins: mapValues(config.plugins, lookup => getModulePath(basedir, lookup)),
+let wasWarned = false
+function warnForNumberedSeverities(result) {
+  if (wasWarned) { return }
+  wasWarned = true
+  result.warn((
+    "Numbered severities (0, 1, 2) have been deprecated, " +
+    "and in 4.0 they will be disabled. "
+  ), {
+    stylelintType: "deprecation",
+    stylelintReference: "http://stylelint.io/?/docs/user-guide/configuration.md",
   })
-}
-
-function getModulePath(basedir, lookup) {
-  try {
-    return resolveFrom(basedir, lookup)
-  } catch (e) {
-    throw configurationError(
-      `Could not find "${lookup}". ` +
-      `Do you need a \`configBasedir\`?`
-    )
-  }
 }