+++ src/rules/block-closing-brace-newline-before/index.js
@@ -1,3 +1,4 @@
+import { startsWith } from "lodash"
 import {
   cssStatementBlockString,
   cssStatementHasBlock,
@@ -11,9 +12,9 @@ import {
 export const ruleName = "block-closing-brace-newline-before"
 
 export const messages = ruleMessages(ruleName, {
-  expectedBefore: () => `Expected newline before "}"`,
-  expectedBeforeMultiLine: () => `Expected newline before "}" of a multi-line block`,
-  rejectedBeforeMultiLine: () => `Unexpected whitespace before "}" of a multi-line block`,
+  expectedBefore: "Expected newline before \"}\"",
+  expectedBeforeMultiLine: "Expected newline before \"}\" of a multi-line block",
+  rejectedBeforeMultiLine: "Unexpected whitespace before \"}\" of a multi-line block",
 })
 
 export default function (expectation) {
@@ -40,50 +41,32 @@ export default function (expectation) {
       const blockIsMultiLine = !isSingleLineString(cssStatementBlockString(statement))
       const after = statement.raw("after")
 
-      if (typeof after === "undefined") { return }
+      if (after === undefined) { return }
 
       // We're really just checking whether a
       // newline *starts* the block's final space -- between
       // the last declaration and the closing brace. We can
       // ignore any other whitespace between them, because that
       // will be checked by the indentation rule.
-      if (after[0] !== "\n" && after.substr(0, 2) !== "\r\n") {
+      if (!startsWith(after, "\n") && !startsWith(after, "\r\n")) {
         if (expectation === "always") {
-          report({
-            message: messages.expectedBefore(),
-            node: statement,
-            index: statement.toString().length - 2,
-            result,
-            ruleName,
-          })
+          complain(messages.expectedBefore)
         } else if (blockIsMultiLine && expectation === "always-multi-line") {
-          report({
-            message: messages.expectedBeforeMultiLine(),
-            node: statement,
-            index: statement.toString().length - 2,
-            result,
-            ruleName,
-          })
+          complain(messages.expectedBeforeMultiLine)
         }
       }
-      if (after !== "") {
-        if (expectation === "never") {
-          report({
-            message: messages.rejectedBefore(),
-            node: statement,
-            index: statement.toString().length - 2,
-            result,
-            ruleName,
-          })
-        } else if (blockIsMultiLine && expectation === "never-multi-line") {
-          report({
-            message: messages.rejectedBeforeMultiLine(),
-            node: statement,
-            index: statement.toString().length - 2,
-            result,
-            ruleName,
-          })
-        }
+      if (after !== "" && blockIsMultiLine && expectation === "never-multi-line") {
+        complain(messages.rejectedBeforeMultiLine)
+      }
+
+      function complain(message) {
+        report({
+          message,
+          result,
+          ruleName,
+          node: statement,
+          index: statement.toString().length - 2,
+        })
       }
     }
   }