+++ src/rules/block-closing-brace-space-after/index.js
@@ -20,7 +20,9 @@ export const messages = ruleMessages(ruleName, {
 })
 
 export default function (expectation) {
+
   const checker = whitespaceChecker("space", expectation, messages)
+
   return function (root, result) {
     const validOptions = validateOptions(result, ruleName, {
       actual: expectation,