+++ src/rules/block-opening-brace-space-before/index.js
@@ -2,7 +2,6 @@ import {
   cssStatementBlockString,
   cssStatementHasBlock,
   cssStatementHasEmptyBlock,
-  cssStatementIsNestingBlock,
   cssStatementStringBeforeBlock,
   report,
   ruleMessages,
@@ -42,8 +41,8 @@ export default function (expectation) {
     root.walkAtRules(check)
 
     function check(statement) {
-      // Return early if blockless, has empty block or is a nesting block
-      if (!cssStatementHasBlock(statement) || cssStatementHasEmptyBlock(statement) || cssStatementIsNestingBlock(statement)) { return }
+      // Return early if blockless or has an empty block
+      if (!cssStatementHasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
 
       const source = cssStatementStringBeforeBlock(statement)