+++ src/rules/declaration-block-semicolon-newline-after/index.js
@@ -16,7 +16,9 @@ export const messages = ruleMessages(ruleName, {
 })
 
 export default function (expectation) {
-  const check = whitespaceChecker("newline", expectation, messages)
+
+  const checker = whitespaceChecker("newline", expectation, messages)
+
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
       actual: expectation,
@@ -42,12 +44,12 @@ export default function (expectation) {
         : nextNode
       if (!nodeToCheck) { return }
 
-      check.afterOneOnly({
+      checker.afterOneOnly({
         source: rawNodeString(nodeToCheck),
         index: -1,
         lineCheckStr: cssStatementBlockString(parentRule),
         err: m => {
-          return report({
+          report({
             message: m,
             node: decl,
             index: decl.toString().length + 1,