+++ src/rules/declaration-block-semicolon-space-before/index.js
@@ -16,7 +16,9 @@ export const messages = ruleMessages(ruleName, {
 })
 
 export default function (expectation) {
-  const check = whitespaceChecker("space", expectation, messages)
+
+  const checker = whitespaceChecker("space", expectation, messages)
+
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
       actual: expectation,
@@ -36,12 +38,12 @@ export default function (expectation) {
 
       const declString = decl.toString()
 
-      check.before({
+      checker.before({
         source: declString,
         index: declString.length,
         lineCheckStr: cssStatementBlockString(parentRule),
         err: m => {
-          return report({
+          report({
             message: m,
             node: decl,
             index: decl.toString().length - 1,