+++ src/rules/max-empty-lines/index.js
@@ -1,3 +1,4 @@
+import { isNumber, repeat } from "lodash"
 import {
   report,
   ruleMessages,
@@ -5,22 +6,27 @@ import {
   validateOptions,
 } from "../../utils"
 
-export const ruleName = "no-multiple-empty-lines"
+export const ruleName = "max-empty-lines"
 
 export const messages = ruleMessages(ruleName, {
   rejected: "Unexpected empty line",
 })
 
-export default function (actual) {
+export default function (max) {
+  const maxAdjacentNewlines = max + 1
+
   return (root, result) => {
-    const validOptions = validateOptions(result, ruleName, { actual })
+    const validOptions = validateOptions(result, ruleName, {
+      actual: max,
+      possible: isNumber,
+    })
     if (!validOptions) { return }
 
     const rootString = root.source.input.css
     styleSearch({ source: rootString, target: "\n", checkComments: true }, match => {
       if (
-        rootString.substr(match.startIndex + 1, 2) === "\n\n"
-        || rootString.substr(match.startIndex + 1, 4) === "\r\n\r\n"
+        rootString.substr(match.startIndex + 1, maxAdjacentNewlines) === repeat("\n", maxAdjacentNewlines)
+        || rootString.substr(match.startIndex + 1, maxAdjacentNewlines * 2) === repeat("\r\n", maxAdjacentNewlines)
       ) {
         report({
           message: messages.rejected,