+++ src/rules/rule-no-duplicate-properties/index.js
@@ -38,6 +38,9 @@ export default function (actual) {
 
         if (cssPropertyIsVariable(prop)) { return }
 
+        // Ignore the src property as commonly duplicated in at-fontface
+        if (prop === "src") { return }
+
         if (decls.indexOf(prop) !== -1) {
           report({
             message: messages.rejected(prop),