+++ src/utils/lineCount.js
@@ -1,9 +1,0 @@
-/**
- * Count the lines in a string.
- *
- * @param {string} source
- * @returns {number} Line count
- */
-export default function (source) {
-  return source.split(/(?:\r\n|\n|\r)/g).length
-}